delete

+++ app/Http/Controllers/Admin/CategoriesController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Category;
+use App\Models\SubCategory;
 
 class CategoriesController extends Controller
 {
@@ -34,6 +35,8 @@ class CategoriesController extends Controller
         $category = new Category; 
         $category->title = request('title');
         $category->description = request('description');
+        $category->is_deleted = 0;
+        $category->status = "active";
         $category->save();
 
         return redirect("/admin/categories")->with('message', 'Category added.');
@@ -58,9 +61,36 @@ class CategoriesController extends Controller
         return redirect("/admin/categories")->with('message', 'Category updated.');
     }
 
+    public function activate($id)
+    {
+        $category = Category::findOrFail($id);
+        $category->update(['status' => "active"]);
+        $category_name = $category->title;
+        return redirect("/admin/categories")->with('message', $category_name." category is now visible.");
+    }
+
+    public function deactivate($id)
+    {
+        $category = Category::findOrFail($id);
+        $category->update(['status' => "inactive"]);
+        $category_name = $category->title;
+        return redirect("/admin/categories")->with('message', $category_name." category is now hidden.");  
+    }
+
     public function destroy($id)
     {
         $category = Category::findOrFail($id);
+        if($category->products->count()){
+            return redirect("/admin/categories")->with('error_message', "Cannot delete, the category has products. Deactivate instead.");  
+        }
+
+        $subcategories = SubCategory::where('category_id',$id)->get();
+        foreach($subcategories as $subcategory){
+            if($subcategory->products->count()){
+                return redirect("/admin/categories")->with('error_message', "Cannot delete, a subcategory of this category has products. Deactivate instead."); 
+            }
+        }
+       
         Category::destroy($category->id);
         return redirect("/admin/categories")->with('message', $category->title." category and its subcategories are now deleted.");  
     }
